pull out RightPanel from ladderPage
