+++ assets/src/components/ladderPage.tsx
@@ -1,28 +1,16 @@
-import * as Sentry from "@sentry/react"
 import { Socket } from "phoenix"
-import React, { ReactElement, useContext, useEffect } from "react"
-import { useHistory } from "react-router-dom"
+import React, { ReactElement, useContext } from "react"
 import RoutesContext from "../contexts/routesContext"
 import { SocketContext } from "../contexts/socketContext"
 import { StateDispatchContext } from "../contexts/stateDispatchContext"
 import { VehiclesByRouteIdProvider } from "../contexts/vehiclesByRouteIdContext"
 import useTimepoints from "../hooks/useTimepoints"
-import useVehicleAndRouteForNotification from "../hooks/useVehicleAndRouteForNotification"
 import useVehicles from "../hooks/useVehicles"
 import { allVehiclesAndGhosts } from "../models/vehiclesByRouteId"
-import {
-  VehicleId,
-  VehicleOrGhost,
-  VehicleOrGhostAndRoute,
-} from "../realtime.d"
+import { VehicleId, VehicleOrGhost } from "../realtime.d"
 import { ByRouteId, Route, RouteId, TimepointsByRouteId } from "../schedule.d"
-import {
-  setNotification,
-  setNotificationIsInactive,
-  setNotificationIsLoading,
-} from "../state"
 import { Notifications } from "./notifications"
-import PropertiesPanel from "./propertiesPanel"
+import RightPanel from "./rightPanel"
 import RouteLadders from "./routeLadders"
 import RoutePicker from "./routePicker"
 
@@ -40,31 +28,9 @@ export const findSelectedVehicleOrGhost = (
   )
 }
 
-const vehicleRoute = (
-  allRoutes: Route[] | null,
-  vehicleOrGhost: VehicleOrGhost | undefined
-): Route | undefined =>
-  (allRoutes || []).find(
-    (route) => route.id === (vehicleOrGhost && vehicleOrGhost.routeId)
-  )
-
-export const chooseVehicleOrGhostForVPP = (
-  vehicleAndRouteForNotification?: VehicleOrGhostAndRoute | null,
-  selectedVehicleOrGhost?: VehicleOrGhost
-): VehicleOrGhost | undefined =>
-  vehicleAndRouteForNotification
-    ? vehicleAndRouteForNotification.vehicleOrGhost
-    : selectedVehicleOrGhost
-
 const LadderPage = (): ReactElement<HTMLDivElement> => {
-  const history = useHistory()
-  if (history) {
-    /* istanbul ignore next */
-    history.listen(() => dispatch(setNotification(undefined)))
-  }
-
-  const [state, dispatch] = useContext(StateDispatchContext)
-  const { selectedRouteIds, selectedVehicleId, selectedNotification } = state
+  const [state] = useContext(StateDispatchContext)
+  const { selectedRouteIds, selectedVehicleId } = state
 
   const routes: Route[] | null = useContext(RoutesContext)
   const timepointsByRouteId: TimepointsByRouteId = useTimepoints(
@@ -77,22 +43,6 @@ const LadderPage = (): ReactElement<HTMLDivElement> => {
     selectedRouteIds
   )
 
-  const vehicleAndRouteForNotification = useVehicleAndRouteForNotification(
-    selectedNotification
-  )
-  useEffect(() => {
-    if (vehicleAndRouteForNotification === null) {
-      dispatch(setNotificationIsInactive())
-    } else if (
-      vehicleAndRouteForNotification === undefined &&
-      selectedNotification
-    ) {
-      dispatch(setNotificationIsLoading(true))
-    } else if (vehicleAndRouteForNotification) {
-      dispatch(setNotificationIsLoading(false))
-    }
-  }, [vehicleAndRouteForNotification, selectedNotification])
-
   const selectedRoutes: Route[] = selectedRouteIds
     .map((routeId) => findRouteById(routes, routeId))
     .filter((route) => route) as Route[]
@@ -101,25 +51,6 @@ const LadderPage = (): ReactElement<HTMLDivElement> => {
     vehiclesByRouteId,
     selectedVehicleId
   )
-
-  const vehicleOrGhostForVPP:
-    | VehicleOrGhost
-    | undefined
-    | null = chooseVehicleOrGhostForVPP(
-    vehicleAndRouteForNotification,
-    selectedVehicleOrGhost
-  )
-
-  const routeForVPP: Route | undefined = vehicleAndRouteForNotification
-    ? vehicleAndRouteForNotification.route
-    : selectedVehicleOrGhost && vehicleRoute(routes, selectedVehicleOrGhost)
-
-  if (vehicleAndRouteForNotification && selectedVehicleOrGhost) {
-    /* istanbul ignore next */
-    Sentry.captureMessage(
-      "vehicleAndRouteForNotification and selectedVehicleOrGhost both set, which should be impossible"
-    )
-  }
 
   return (
     <div className="m-ladder-page">
@@ -133,13 +64,7 @@ const LadderPage = (): ReactElement<HTMLDivElement> => {
             timepointsByRouteId={timepointsByRouteId}
             selectedVehicleId={selectedVehicleId}
           />
-
-          {vehicleOrGhostForVPP && routeForVPP && (
-            <PropertiesPanel
-              selectedVehicleOrGhost={vehicleOrGhostForVPP}
-              route={routeForVPP}
-            />
-          )}
+          <RightPanel selectedVehicleOrGhost={selectedVehicleOrGhost} />
         </>
       </VehiclesByRouteIdProvider>
     </div>
